% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluserting.R
\name{cluster_cells}
\alias{cluster_cells}
\title{Perform cell clustering based on scDHA method}
\usage{
cluster_cells(ScRNA_filtered, Normalize = TRUE, k=NULL, n=5000)
}
\arguments{
\item{ScRNA_filtered}{ScRNA-seq data set generated by filter_ScRNA function}

\item{Normalize}{Boolean parameter whether to apply log10 normalization for the data or not}

\item{k}{Number of clusters if there is a prior knowledge about that}

\item{n}{Number of genes to keep after feature selection step}
}
\value{
a vector that contains the cell labels
}
\description{
Cluster cells based on scDHA methods to get cluster labels
}
\author{
Mohamed Soudy \email{Mohmedsoudy2009@gmail.com}
}
