% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluserting.R
\name{evaluate_clustering}
\alias{evaluate_clustering}
\title{Evaluate the clustering if you have the original labels}
\usage{
evaluate_clustering(cluster_labels, original_labels)
}
\arguments{
\item{cluster_labels}{Cluster labels generated by cluster_cells functions or user-defined}

\item{original_labels}{Original labels of the ScRNA-seq data}
}
\value{
ARI of clustering 'a value between 0 and 1' 1 indicates best clustering
}
\description{
Evaluate the clustering if you have the original labels
}
\examples{
evaluate_clustering(c(1,1,1,1,2,2,3,3), c(1,1,1,1,3,3,3,2))

}
\author{
Mohamed Soudy \email{Mohmedsoudy2009@gmail.com}
}
