% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{filter_ScRNA}
\alias{filter_ScRNA}
\title{Remove genes which are not expressed in at least one cell}
\usage{
filter_ScRNA(ScRNA_mat)
}
\arguments{
\item{ScRNA_mat}{ScRNA-Seq matrix where genes are in rows and cells are in columns}
}
\value{
Filtered ScRNA-seq matrix that contains genes that are at least expressed in one cell type
}
\description{
Filter out the genes that are not expressed in at least one cell type
}
\author{
Mohamed Soudy \email{Mohmedsoudy2009@gmail.com}
}
