% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{prepare_dataset}
\alias{prepare_dataset}
\title{Prepare the data set for the imputation}
\usage{
prepare_dataset(filtered_data, cluster_labels)
}
\arguments{
\item{filtered_data}{ScRNA-seq data set generate by filter_ScRNA function}

\item{cluster_labels}{cell labels}
}
\value{
a data frame that contains the processed ScRNA-seq data
}
\description{
This function aims to get the indices where all genes are zeros or genes that is expressed on only one sample per cell
}
\author{
Mohamed Soudy \email{Mohmedsoudy2009@gmail.com}
}
