% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScRNAIMM.R
\name{run_pipeline}
\alias{run_pipeline}
\title{Run the main pipeline for ScRNAIMM}
\usage{
run_pipeline(ScRNA,label=NULL,k=NULL,cells=TRUE,genes=TRUE,outdir=NULL,dataset=NULL)
}
\arguments{
\item{ScRNA}{ScRNA-seq data set generated by prepare_dataset function}

\item{label}{Prior knowledge about cluster labels if NULL, will use our clustering function}

\item{k}{Prior knowledge about number of clusters if NULL, will use our clustering function}

\item{cells}{Boolean whether to do the imputation based on cell clustering or not}

\item{genes}{Boolean whether to do the imputation based on genes or not}

\item{outdir}{Path to output directory to write the imputed data}

\item{dataset}{Name of the data set to be the name of the output directory}
}
\value{
a data frame with the imputed values
}
\description{
Run the main pipeline for ScRNAIMM
}
\author{
Mohamed Soudy \email{Mohmedsoudy2009@gmail.com}
}
