% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScRNAIMM.R
\name{ScRNA_imp_MM}
\alias{ScRNA_imp_MM}
\title{Perform ScRNA-seq imputation using mean/Median}
\usage{
ScRNA_imp_MM(ScRNA_filtered, cluster_labels = NULL, cells = TRUE, genes = FALSE)
}
\arguments{
\item{ScRNA_filtered}{ScRNA-seq data set generated by prepare_dataset function}

\item{cluster_labels}{Cluster labels generated by cluster_cells function or user-defined}

\item{cells}{Boolean whether to do the imputation based on cell clustering or not}

\item{genes}{Boolean whether to do the imputation based on genes or not}
}
\value{
a data frame with the imputed values
}
\description{
Perform ScRNA-seq imputation using mean/Median
}
\author{
Mohamed Soudy \email{Mohmedsoudy2009@gmail.com}
}
