% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScRNAIMM.R
\name{scRNA_MMI}
\alias{scRNA_MMI}
\title{Performs a distribution check for the data}
\usage{
scRNA_MMI(scRNA_dataset)
}
\arguments{
\item{scRNA_dataset}{ScRNA-seq data set}
}
\value{
ScRNA-Seq Imputed data set
}
\description{
Performs a distribution check for the data
}
\author{
Mohamed Soudy \email{Mohmedsoudy2009@gmail.com}
}
